Added Search and Item Selection

+++ src/test/java/testBase/BaseClass.java
@@ -3,6 +3,7 @@ package testBase;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.net.URL;
 import java.text.SimpleDateFormat;
 import java.time.Duration;
 import java.util.Date;
@@ -12,11 +13,14 @@ import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;  //log4j2
 import org.openqa.selenium.OutputType;
+import org.openqa.selenium.Platform;
 import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;	 //log4j2
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Parameters;
@@ -36,21 +40,57 @@ public class BaseClass {
 		p=new Properties();
 		p.load(file);
 		
-		
 		logger=LogManager.getLogger(this.getClass()); //log4j2
 		
-		switch(br.toLowerCase())
+		if(p.getProperty("execution_env").equalsIgnoreCase("remote"))
 		{
-		case "chrome": driver=new ChromeDriver(); break;
-		case "edge": driver=new EdgeDriver(); break;
-		case "Firefox": driver=new FirefoxDriver(); break;
-		default : System.out.println("Invalid Browser"); return;
-		}
+			DesiredCapabilities cap=new DesiredCapabilities();
+			
+			if(os.equalsIgnoreCase("windows"))
+				{
+				cap.setPlatform(Platform.WIN11);
+			}
+			else if(os.equalsIgnoreCase("mac"))
+			{
+				cap.setPlatform(Platform.MAC);
+			}
+			else
+			{
+				System.out.println("No matching os");
+				return;
+			}
+			
+			switch(br.toLowerCase())
+			{
+			case "chrome": cap.setBrowserName("chrome"); break;
+			case "edge": cap.setBrowserName("MicrosoftEdge"); break;
+			default: System.out.println("No matching browser"); return;
+			}
+			
+			driver = new RemoteWebDriver(new URL("http://localhost:4444/wd/hub"), cap);
+			driver.manage().deleteAllCookies();
+			driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+			driver.get(p.getProperty("appUrl")); //reading url from properties file
+			driver.manage().window().maximize();
+
 		
-		driver.manage().deleteAllCookies();
-		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
-		driver.get(p.getProperty("appUrl")); //reading url from properties file
-		driver.manage().window().maximize();
+		}
+		if(p.getProperty("execution_env").equalsIgnoreCase("local"))
+		{
+			switch(br.toLowerCase())
+			{
+			case "chrome": driver=new ChromeDriver(); break;
+			case "edge": driver=new EdgeDriver(); break;
+			case "Firefox": driver=new FirefoxDriver(); break;
+			default : System.out.println("Invalid Browser"); return;
+			}
+			
+			driver.manage().deleteAllCookies();
+			driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+			driver.get(p.getProperty("appUrl")); //reading url from properties file
+			driver.manage().window().maximize();
+		}
+
 	}
 				
 		@AfterClass(groups= {"Sanity","Regression","Master"})
